CustomDomain improvements

+++ bin/execute_custom_domain.sh
@@ -1,13 +1,37 @@
+#!/bin/bash
 for file in var/file_queues/custom_domain_to_remove/*
-do
-    CUSTOM_URL=`cat $file`
-    rm /etc/nginx/sites-available/$CUSTOM_URL /etc/nginx/sites-enabled/$CUSTOM_URL
-    rm $file
+do    
+    if [ -e $file ] 
+    then
+        CUSTOM_URL=`cat $file`
+        echo "Removing $file with content $CUSTOM_URL"
+        if [ -n "$CUSTOM_URL" ]; then
+            echo "Remove previous certificates if exists"
+            if [ -e /etc/letsencrypt/live/$CUSTOM_URL/cert.pem ]; then
+                certbot revoke -n --cert-path /etc/letsencrypt/live/$CUSTOM_URL/cert.pem            
+            fi 
+            echo "Delete previous nginx conf"
+            rm /etc/nginx/sites-available/$CUSTOM_URL /etc/nginx/sites-enabled/$CUSTOM_URL
+            rm $file       
+            echo "Reload Nginx" 
+            service nginx reload       
+        fi
+    fi
 done
 
 for file in var/file_queues/custom_domain_to_configure/*
 do
-    arguments=`cat $file`
-    sh bin/configure_custom_domain.sh $arguments
-    rm $file
+    if [ -e $file ] 
+    then
+        arguments=`cat $file`
+        echo "configure $file with content $arguments"
+
+        if sh bin/configure_custom_domain.sh $arguments
+        then
+            echo "Successfully configured $file, the file queue can now be deleted"
+            rm $file
+        else
+            echo "Problem while configuring $file"
+        fi    
+    fi
 done
